requirements.txt
Flask==2.1.2
requests==2.28.1
requests-oauthlib==1.3.1
python-dotenv==0.19.2
